Add option to not set xml:base when serialising URL content to string.
When grabbing CellML models we make use of xml:base attribute to resolve relative import URLs, but now we are using the csim::Model::serialiseUrlToString method to grab SED-ML
documents in get-simulator - and libSEDML says the SED-ML is invalid if an xml:base is set. So have added an option to not set the xml:base attribute to this convenience method.